% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{AJ_AlmacoJack_OM}
\alias{AJ_AlmacoJack_OM}
\title{AJ_AlmacoJack_OM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
AJ_AlmacoJack_OM
}
\description{
An object of class OM
}
\keyword{datasets}
