% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Arrowtooth_Flounder_BC_SS}
\alias{Arrowtooth_Flounder_BC_SS}
\title{Arrowtooth_Flounder_BC_SS OM}
\format{An object of class \code{OM} of length 1.}
\usage{
Arrowtooth_Flounder_BC_SS
}
\description{
An object of class OM (from iSCAM using iSCAM2DLM)
 DFO. Author: DFO.
}
\keyword{datasets}
