% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Canary_Rockfish_BC_CanaryRockfishBC_OM}
\alias{Canary_Rockfish_BC_CanaryRockfishBC_OM}
\title{Canary_Rockfish_BC_CanaryRockfishBC_OM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
Canary_Rockfish_BC_CanaryRockfishBC_OM
}
\description{
An object of class OM
}
\keyword{datasets}
