% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Darnley_Bay_Char_ArcticCharr_OM}
\alias{Darnley_Bay_Char_ArcticCharr_OM}
\title{Darnley_Bay_Char_ArcticCharr_OM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
Darnley_Bay_Char_ArcticCharr_OM
}
\description{
An object of class OM
}
\keyword{datasets}
