% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{LAJ_LesserAmberjack_OM}
\alias{LAJ_LesserAmberjack_OM}
\title{LAJ_LesserAmberjack_OM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
LAJ_LesserAmberjack_OM
}
\description{
An object of class OM
}
\keyword{datasets}
