% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{LS_LaneSnapper_OM}
\alias{LS_LaneSnapper_OM}
\title{LS_LaneSnapper_OM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
LS_LaneSnapper_OM
}
\description{
An object of class OM
}
\keyword{datasets}
