% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{RD_RedDrum_OM}
\alias{RD_RedDrum_OM}
\title{RD_RedDrum_OM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
RD_RedDrum_OM
}
\description{
An object of class OM
}
\keyword{datasets}
