% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Red_snapper_GOM_SS}
\alias{Red_snapper_GOM_SS}
\title{Red_snapper_GOM_SS OM}
\format{An object of class \code{OM} of length 1.}
\usage{
Red_snapper_GOM_SS
}
\description{
An object of class OM (from SS using SS2DLM)
 SEDAR 31. Author: SEDAR 31.
}
\keyword{datasets}
