% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{San_Francisco_Bay_herring_HerringSFB_OM}
\alias{San_Francisco_Bay_herring_HerringSFB_OM}
\title{San_Francisco_Bay_herring_HerringSFB_OM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
San_Francisco_Bay_herring_HerringSFB_OM
}
\description{
An object of class OM
}
\keyword{datasets}
