% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{SG_SnowyGrouper_OM}
\alias{SG_SnowyGrouper_OM}
\title{SG_SnowyGrouper_OM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
SG_SnowyGrouper_OM
}
\description{
An object of class OM
}
\keyword{datasets}
