% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{SH_SpeckledHind_OM}
\alias{SH_SpeckledHind_OM}
\title{SH_SpeckledHind_OM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
SH_SpeckledHind_OM
}
\description{
An object of class OM
}
\keyword{datasets}
