% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{WM_Wenchman_OM}
\alias{WM_Wenchman_OM}
\title{WM_Wenchman_OM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
WM_Wenchman_OM
}
\description{
An object of class OM
}
\keyword{datasets}
