% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{YM_YellowmouthGrouper_OM}
\alias{YM_YellowmouthGrouper_OM}
\title{YM_YellowmouthGrouper_OM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
YM_YellowmouthGrouper_OM
}
\description{
An object of class OM
}
\keyword{datasets}
